fix: dont retry 401 responses on GraphQL requests
This updates the bottleneck "failed" handler to not retry GraphQL
requests that receive a 401 response. The old behavior was to retry all
GraphQL error responses, which caused the retry to be throttled for an
hour, due to 401 Unauthorized responses including an
`x-ratelimit-remaining` of `0` and a `x-ratelimit-reset` of now + 3600
seconds. The main issue was that the request wasn't being ratelimited,
but the ratelimit headers caused the plugin to wait.

In testing this out, I noticed that GraphQL responses can also receive a
403 Forbidden response when hitting the abuse rate limit. This commit
keeps the functionality for retrying 403 responses for GraphQL requests,
since they signify abuse rate limiting, where 401 responses indicate bad
credentials.